Split expected composite symbol tests
